pyproject.toml
[tool.poetry]
name = "ytdl_api_server"
version = "0.1.0"
description = "Sqlite-based state tracking API server for ytdl downloads"
authors = ["Kim van Wyk <kimv@sahomeloans.com>"]

[tool.poetry.dependencies]
python = "^3.7"
black = "^20.8b1"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
sqlalchemy = "^1.3.23"
attrs = "^20.3.0"

[tool.poetry.dev-dependencies]
ptpython = "^3.0.13"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
